+++ src/geometric_object/bvh_node.rs
@@ -50,7 +50,7 @@ impl Geometry for BvhNode {
         Aabb::new(self.get_min_point(), self.get_max_point())
     }
 
-    fn get_samples(&self, _sample_points_sqrt: usize) -> Vec<Point3<f64>> {
+    fn get_samples(&self, _sample_points_sqrt: u8) -> Vec<Point3<f64>> {
         vec![]
     }
