+++ src/geometric_object/sphere.rs
@@ -96,7 +96,7 @@ impl Geometry for Sphere {
         Aabb::new(self.get_min_point(), self.get_max_point())
     }
 
-    fn get_samples(&self, _sample_points_sqrt: usize) -> Vec<Point3<f64>> {
+    fn get_samples(&self, _sample_points_sqrt: u8) -> Vec<Point3<f64>> {
         vec![]
     }
