+++ src/sampler.rs
@@ -6,14 +6,14 @@ use std::f64::consts::FRAC_PI_4;
 use crate::geometric_object::Triangle;
 use crate::model::Vec3;
 
-pub fn get_square_sampler(n: usize) -> impl Iterator<Item = Point2<f64>> {
+pub fn get_square_sampler(n: u8) -> impl Iterator<Item = Point2<f64>> {
     thread_rng()
         .sample_iter(&Standard)
-        .take(n * n)
+        .take((n * n) as usize)
         .map(|(i, j)| Point2::new(i, j))
 }
 
-pub fn get_triangle_sampler(n: usize, t: &Triangle) -> impl Iterator<Item = Point3<f64>> {
+pub fn get_triangle_sampler(n: u8, t: &Triangle) -> impl Iterator<Item = Point3<f64>> {
     let (x, y, z) = (t.x, t.y, t.z);
     get_square_sampler(n).map(move |p| {
         let mut a = p.x;
@@ -26,7 +26,7 @@ pub fn get_triangle_sampler(n: usize, t: &Triangle) -> impl Iterator<Item = Poin
     })
 }
 
-pub fn get_hemisphere_sampler(n: usize) -> impl Iterator<Item = Vec3> {
+pub fn get_hemisphere_sampler(n: u8) -> impl Iterator<Item = Vec3> {
     get_square_sampler(n).map(|p| {
         let e = 1.0;
         let phi = 2.0 * std::f64::consts::PI * p.x;
@@ -38,7 +38,7 @@ pub fn get_hemisphere_sampler(n: usize) -> impl Iterator<Item = Vec3> {
     })
 }
 
-pub fn get_disk_sampler(n: usize) -> impl Iterator<Item = (Point2<f64>, Point2<f64>)> {
+pub fn get_disk_sampler(n: u8) -> impl Iterator<Item = (Point2<f64>, Point2<f64>)> {
     get_square_sampler(n).map(|p| {
         let spx = 2.0 * p.x - 1.0;
         let spy = 2.0 * p.y - 1.0;
